Renamed Resources annotation to ConcordionResources. Since there is a javax.annotation.Resources class, we felt that this would confuse some users who aren't proficient Java developers. Also some users may be using CDI, and would need to specify the full package name for one or other of the Resources annotations.
