+++ src/test-dummies/resources/resources.css
@@ -1,1 +1,1 @@
-/* Here is a CSS file (Used by Resources test) */
+/* Here is a CSS file (Used by ConcordionResources test) */
