+++ src/test/resources/spec/concordion/annotation/Annotation.html
@@ -6,7 +6,7 @@
 	<h2>Customise the specification</h2>
 	
 	<p>
-		The <a concordion:run="concordion" href="Resources.html">Resources Annotation</a> can be used to apply a new theme to 
+		The <a concordion:run="concordion" href="ConcordionResources.html">ConcordionResources Annotation</a> can be used to apply a new theme to 
 		your specifications, tweak the existing styling, or add new css, javascript, images, or other resources.
 	 </p>
