+++ src/test/resources/spec/concordion/extension/ResourceExtension.html
@@ -6,7 +6,13 @@
 
     <p>
        Resources can be added to the Concordion output folder using an <a href="Extension.html">extension</a>. 
+       This is useful if your extension requires resources, such as applying additional styling. 
     </p>
+    <p>
+       Since Concordion 2.0.0, you can also add resources using the <a href="../annotation/ConcordionResources.html">ConcordionResources</a> annotation. This is a simpler mechanism if
+       your goal is just to add resources, without introducing additional functionality.
+    </p>
+    <p>This specification describes the first case, how to add resources using an extension.</p>
 
     <h3>Static Resources</h3>
     <p>
